Added check for anonymous (not logged in) user
Logged in user saves ingredients, not logged in user can still search based on ingredients added.

+++ recipes/views.py
@@ -19,13 +19,14 @@ class SearchRecipes(View):
         """
         print("InPOST")
         ingredient_ids = json.loads(request.POST['ingredient_ids'])
-        useringredients = list(UserProfile.get_or_create_profile(self.request.user).ingredients.values_list('id', flat=True))
-        ingredientstoremove = list(set(useringredients) - set(ingredient_ids))
-        ingredientstoadd = list(set(ingredient_ids) - set(useringredients))
-        for ingredient in ingredientstoremove:
-            self.request.user.userprofile.delete_user_ingredient(ingredient)
-        if len(ingredientstoadd) > 0:
-            self.request.user.userprofile.add_user_ingredients(ingredientstoadd)
+        if not self.request.user.is_anonymous():
+            useringredients = list(UserProfile.get_or_create_profile(self.request.user).ingredients.values_list('id', flat=True))
+            ingredientstoremove = list(set(useringredients) - set(ingredient_ids))
+            ingredientstoadd = list(set(ingredient_ids) - set(useringredients))
+            for ingredient in ingredientstoremove:
+                self.request.user.userprofile.delete_user_ingredient(ingredient)
+            if len(ingredientstoadd) > 0:
+                self.request.user.userprofile.add_user_ingredients(ingredientstoadd)
         context = {
             'results': Recipe.get_recipes_by_ingredients(ingredient_ids),
             'parameters': []
